fix dependancy cycle eslint errors
